% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{antimoda}
\alias{antimoda}
\title{Get anti-mode}
\usage{
antimoda(x)
}
\arguments{
\item{x}{A categorical variable}
}
\value{
The anti-mode (least repeated value)
}
\description{
Returns the least repeated value
}
