% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_quality.R
\name{bivariate_outliers}
\alias{bivariate_outliers}
\title{Check for bivariate outliers}
\usage{
bivariate_outliers(x, threshold_d = 10, threshold_b = 1.5)
}
\arguments{
\item{x}{A data.frame object}

\item{threshold_d}{Threshold for the case of two continuous variables}

\item{threshold_b}{Threshold for the case of one continuous and one categorical variable}
}
\value{
A data frame with all the observations considered as bivariate outliers
}
\description{
Checks for bivariate outliers in a data.frame
}
\examples{
bivariate_outliers(iris)
}
