% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{clickR-defunct}
\alias{clickR-defunct}
\title{Defunct functions in clickR}
\description{
These functions are defunct, please use the package repmod to find their updated versions.
}
\details{
\itemize{
 \item \code{\link{rob.pvals}}: This function is defunct, and will
 be removed in the next version of this package.
 \item \code{\link{rob.ci}}: This function is defunct, and will
 be removed in the next version of this package.
 \item \code{\link{make_word_table}}: This function is defunct, and will
 be removed in the next version of this package.
 \item \code{\link{make_latex_table}}: This function is defunct, and will
 be removed in the next version of this package.
 \item \code{\link{make_csv_table}}: This function is defunct, and will
 be removed in the next version of this package.
 \item \code{\link{make_table}}: This function is defunct, and will
 be removed in the next version of this package.
 \item \code{\link{VarCorr}}: This function is defunct, and will
 be removed in the next version of this package.
 \item \code{\link{plot.reportmodel}}: This function is defunct, and will
 be removed in the next version of this package.
 \item \code{\link{coefplot}}: This function is defunct, and will
 be removed in the next version of this package.
}
}
