% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{cluster_var}
\alias{cluster_var}
\title{Clustering of variables}
\usage{
cluster_var(x, margins = c(8, 1))
}
\arguments{
\item{x}{A data.frame}

\item{margins}{Margins for the plot}
}
\value{
A heatmap with the variable associations
}
\description{
Displays associations between variables in a data.frame in a heatmap with clustering
}
\examples{
cluster_var(iris)
cluster_var(mtcars)
}
