% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{descriptive}
\alias{descriptive}
\title{Detailed summary of the data}
\usage{
descriptive(x, z = 3, ignore.na = TRUE, by = NULL)
}
\arguments{
\item{x}{A data.frame}

\item{z}{Number of decimal places}

\item{ignore.na}{If TRUE NA values will not count for relative frequencies calculations}

\item{by}{Factor variable definining groups for the summary}
}
\value{
Summary of the data
}
\description{
Creates a detailed summary of the data
}
\examples{
descriptive(iris)
descriptive(iris, by="Species")
}
