% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{descriptivo}
\alias{descriptivo}
\title{Defunct function for creating data summaries}
\usage{
descriptivo(x)
}
\arguments{
\item{x}{A data.frame}
}
\value{
Nothing, the function is defunct. Use descriptive() instead.
}
\description{
Creates a detailed summary of the data
}
