% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_quality.R
\name{extreme_values}
\alias{extreme_values}
\title{Extreme values from a numeric vector}
\usage{
extreme_values(x, n = 5, id = NULL)
}
\arguments{
\item{x}{A vector}

\item{n}{Number of extreme values to return}

\item{id}{ID column to reference the found extreme values}
}
\value{
A matrix with the lowest and highest values from a vector
}
\description{
Returns the nth lowest and highest values from a vector
}
