% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{fix.all}
\alias{fix.all}
\title{fix.all}
\usage{
fix.all(x, track = TRUE)
}
\arguments{
\item{x}{A data.frame}

\item{track}{Track changes?}
}
\description{
Tries to fix all problems in the data.frame
}
