% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_functions.R
\name{fix.concat}
\alias{fix.concat}
\title{fix.concat}
\usage{
fix.concat(x, varname, sep = ", |; | ", track = TRUE)
}
\arguments{
\item{x}{A data.frame}

\item{varname}{Variable name}

\item{sep}{Separator for the different values}

\item{track}{Track changes?}
}
\description{
Fixes concatenated values in a variable
}
\examples{
mydata <- data.frame(concat=c("a", "b", "a b" , "a b, c", "a; c"),
numeric = c(1, 2, 3, 4, 5))
fix.concat(mydata, "concat")
}
