% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{fix.NA}
\alias{fix.NA}
\title{fix.NA}
\usage{
fix.NA(
  x,
  na.strings = c("^$", "^ $", "^\\\\?$", "^-$", "^\\\\.$", "^NaN$", "^NULL$"),
  track = TRUE
)
}
\arguments{
\item{x}{A data.frame}

\item{na.strings}{Strings to be considered NA}

\item{track}{Track changes?}
}
\description{
Fixes miscoded missing values
}
\examples{
mydata <- data.frame(prueba = c("", NA, "A", 4, " ", "?", "-", "+"),
casa = c("", 1, 2, 3, 4, " ", 6, 7))
fix.NA(mydata)
}
