% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{forge}
\alias{forge}
\title{Forge}
\usage{
forge(data, affixes, force.fixed = NULL)
}
\arguments{
\item{data}{data.frame}

\item{affixes}{Affixes for repeated measures}

\item{force.fixed}{Variables with matching affix to be excluded}
}
\description{
Reshapes a data frame from wide to long format
}
