% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_functions.R
\name{fxd}
\alias{fxd}
\title{Internal function to fix_dates}
\usage{
fxd(d, use.probs = TRUE)
}
\arguments{
\item{d}{A character vector}

\item{use.probs}{Solve ambiguities by similarity to the most frequent formats}
}
\description{
Function to format dates
}
