% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{GK_assoc}
\alias{GK_assoc}
\title{Computes Goodman and Kruskal's tau}
\usage{
GK_assoc(x, y)
}
\arguments{
\item{x}{A categorical variable}

\item{y}{A categorical variable}
}
\value{
Goodman and Kruskal's tau
}
\description{
Returns Goodman and Kruskal's tay measure of association between two categorical variables
}
