% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{good2go}
\alias{good2go}
\title{Good to go}
\usage{
good2go(path = getwd(), info = TRUE, load = TRUE)
}
\arguments{
\item{path}{Path where the scripts are located}

\item{info}{List the libraries found?}

\item{load}{Should the libraries found be loaded?}
}
\description{
Loads all libraries used in scripts inside the selected path
}
