% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_quality.R
\name{\%<NA\%}
\alias{\%<NA\%}
\title{less & NA}
\usage{
x \%<NA\% y
}
\arguments{
\item{x}{Vector for the left side of the operator}

\item{y}{A Scalar or vector of the same length as x for the right side of the operator}
}
\value{
A logical vector of the same length as x
}
\description{
'<' operator where NA values return FALSE
}
