% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{ipboxplot}
\alias{ipboxplot}
\title{Improved boxplot}
\usage{
ipboxplot(formula, boxwex = 0.6, ...)
}
\arguments{
\item{formula}{Formula for the boxplot}

\item{boxwex}{Width of the boxes}

\item{...}{further arguments passed to beeswarm()}
}
\description{
Creates an improved boxplot with individual data points
}
\examples{
ipboxplot(Sepal.Length ~ Species, data=iris)
ipboxplot(mpg ~ gear, data=mtcars)
}
