% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{kill.factors}
\alias{kill.factors}
\title{Kill factors}
\usage{
kill.factors(dat, k = 10)
}
\arguments{
\item{dat}{A data.frame}

\item{k}{Maximum number of levels for factors}
}
\description{
Changes factor variables to character
}
\examples{
d <- data.frame(Letters=letters[1:20], Nums=1:20)
d$Letters
d <- kill.factors(d)
d$Letters
}
