% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_model.R
\name{make_latex_table}
\alias{make_latex_table}
\title{Defunct functions in package \sQuote{clickR}}
\usage{
make_latex_table(...)
}
\description{
Exports a table to latex. This function has been
removed from the package. Updated functions are available in \sQuote{repmod} package.
}
\details{
Export a table to latex
}
