% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_model.R
\name{make_table}
\alias{make_table}
\title{Make a table from report}
\usage{
make_table(x, file, type, info = NULL, ...)
}
\arguments{
\item{x}{A data.frame object}

\item{file}{Name of the file}

\item{type}{Type of file}

\item{info}{Footer for the table}

\item{...}{Additional parameters passed to make_word_table}
}
\value{
Creates a file with the table
}
\description{
Auxiliary function to create tables
}
