% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_model.R
\name{make_word_table}
\alias{make_word_table}
\title{Deprecated functions in package \sQuote{clickR}}
\usage{
make_word_table(x, file, info = NULL, use.rownames = TRUE)
}
\arguments{
\item{x}{A data.frame object}

\item{file}{Name of the file}

\item{info}{Footer for the table}

\item{use.rownames}{Should row names be added to the output?}
}
\value{
Creates a word file with the table
}
\description{
Exports a table to Word. This function will be
defunct in the next version. Updated functions are available in \sQuote{repmod} package.
}
\details{
Export a table to word
}
