% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_quality.R
\name{may.numeric}
\alias{may.numeric}
\title{Checks if each value might be numeric}
\usage{
may.numeric(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A logical vector
}
\description{
Checks if each value from a vector might be numeric
}
