% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{moda_cont}
\alias{moda_cont}
\title{Estimates number of modes}
\usage{
moda_cont(x)
}
\arguments{
\item{x}{A numeric variable}
}
\value{
Estimated number of modes.
}
\description{
Estimates the number of modes
}
