% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{mtapply}
\alias{mtapply}
\title{Multiple tapply}
\usage{
mtapply(x, group, fun)
}
\arguments{
\item{x}{A data.frame}

\item{group}{Grouping variable}

\item{fun}{Function to apply by group}
}
\description{
Modification of the tapply function to use with data.frames
}
\examples{
mtapply(mtcars, mtcars$gear, mean)
}
