% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtcars_messy-data.R
\docType{data}
\name{mtcars_messy}
\alias{mtcars_messy}
\title{Messy Motor Trend Car Road Tests Dataset}
\format{
A data frame with 32 observations and 13 variables
}
\source{
\code{datasets} package
}
\usage{
data(mtcars_messy)
}
\description{
Modified version of the mtcars dataset with different types of errors
in the data. The dataset has 13 variables and 32 observations.
}
\examples{
data(mtcars_messy)
descriptive(mtcars_messy)
}
\references{
Henderson and Velleman (1981), Building multiple regression
models interactively. Biometrics, 37, 391–411.
}
\keyword{datasets}
