% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{nearest}
\alias{nearest}
\title{Internal function for descriptive()}
\usage{
nearest(x, to = seq(0, 1, length.out = 30))
}
\arguments{
\item{x}{A numeric value between 0-1}

\item{to}{Range of reference values}
}
\value{
The nearest position to the input value
}
\description{
Finds positions for substitution of characters in Distribution column
}
