% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{nice_names}
\alias{nice_names}
\title{Nice names}
\usage{
nice_names(dat)
}
\arguments{
\item{dat}{A data.frame}
}
\description{
Changes names of a data frame to ease work with them
}
\examples{
d <- data.frame('Variable 1'=NA, '\% Response'=NA, ' Variable     3'=NA,check.names=FALSE)
names(d)
names(nice_names(d))
}
