% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_quality.R
\name{numeros}
\alias{numeros}
\title{Brute numeric coercion}
\usage{
numeros(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A numeric vector
}
\description{
If possible, coerces values from a vector to numeric
}
