% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{outliers}
\alias{outliers}
\title{outliers}
\usage{
outliers(x, threshold = 1.5)
}
\arguments{
\item{x}{A vector}

\item{threshold}{Threshold (as multiple of the IQR) to consider an observation as outlier}
}
\description{
Function for detecting outliers based on the boxplot method
}
\examples{
outliers(iris$Petal.Length)
outliers(airquality$Ozone)
}
