% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{peek}
\alias{peek}
\title{Peek}
\usage{
peek(x, n = 10, which = 1:ncol(x))
}
\arguments{
\item{x}{A data.frame}

\item{n}{Number of rows to include in output}

\item{which}{Columns to include in output}
}
\description{
Takes a peek into a data.frame returning a concise visualization about it
}
\examples{
peek(iris)
}
