% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefplot.R
\name{plot.reportmodel}
\alias{plot.reportmodel}
\title{Coefplot for reportmodel objects}
\usage{
\method{plot}{reportmodel}(x, ...)
}
\arguments{
\item{x}{A reportmodel object}

\item{...}{Further arguments passed to coefplot}
}
\description{
Creates a coefplot from the reportmodel object
}
