% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{prop_may}
\alias{prop_may}
\title{Gets proportion of most repeated value}
\usage{
prop_may(x, ignore.na = TRUE)
}
\arguments{
\item{x}{A categorical variable}

\item{ignore.na}{Should NA values be ignored for computing proportions?}
}
\value{
A proportion
}
\description{
Returns the proportion for the most repeated value
}
