% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{prop_min}
\alias{prop_min}
\title{Gets proportion of least repeated value}
\usage{
prop_min(x, ignore.na = TRUE)
}
\arguments{
\item{x}{A categorical variable}

\item{ignore.na}{Should NA values be ignored for computing proportions?}
}
\value{
A proportion
}
\description{
Returns the proportion for the least repeated value
}
