% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_functions.R
\name{remove_empty}
\alias{remove_empty}
\title{remove_empty}
\usage{
remove_empty(x, track = TRUE)
}
\arguments{
\item{x}{A data.frame}

\item{track}{Track changes?}
}
\description{
Removes empty rows or columns from data.frames
}
\examples{
mydata <- data.frame(a = c(NA, NA, NA, NA, NA), b = c(1, NA, 3, 4, 5),
c=c(NA, NA, NA, NA, NA), d=c(4, NA, 5, 6, 3))
remove_empty(mydata)
}
