% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_model.R
\name{report.clmm}
\alias{report.clmm}
\title{Report from ordinal mixed model}
\usage{
\method{report}{clmm}(x, file = NULL, type = "word", digits = 3,
  digitspvals = 3, font = ifelse(Sys.info()["sysname"] == "Windows",
  "Arial", "Helvetica")[[1]], pointsize = 11, ...)
}
\arguments{
\item{x}{An ordinal model object}

\item{file}{Name of the file to export the table}

\item{type}{Format of the file}

\item{digits}{Number of decimals}

\item{digitspvals}{Number of decimals for p-values}

\item{font}{Font to use if type="word"}

\item{pointsize}{Pointsize to use if type="word"}

\item{...}{Further arguments passed to make_table}
}
\value{
A data frame with the report table
}
\description{
Creates a report table from an ordinal mixed model
}
