% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_model.R
\name{report.coxph}
\alias{report.coxph}
\title{Defunct functions in package \sQuote{clickR}}
\usage{
report.coxph(...)
}
\description{
Creates a report table from a cox model. This function has been
removed from the package. Updated functions are available in \sQuote{repmod} package.
}
\details{
Report from cox regression model
}
