% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_model.R
\name{report.default}
\alias{report.default}
\title{Defunct functions in package \sQuote{clickR}}
\usage{
report.default(...)
}
\description{
This is a default function for calling summary(x) on non-implemented classes. This function has been
removed from the package. Updated functions are available in \sQuote{repmod} package.
}
\details{
Default function for report
}
