% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_model.R
\name{report.factor}
\alias{report.factor}
\title{Report from categorical variable}
\usage{
\method{report}{factor}(x, ...)
}
\arguments{
\item{x}{A categorical variable}

\item{...}{Further arguments passed to make_table}
}
\value{
A data frame with the report table
}
\description{
Creates a report table
}
