% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_model.R
\name{report.glm}
\alias{report.glm}
\title{Defunct functions in package \sQuote{clickR}}
\usage{
report.glm(...)
}
\description{
Creates a report table from a generalized linear model. This function has been
removed from the package. Updated functions are available in \sQuote{repmod} package.
}
\details{
Report from generalized linear model
}
