% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_model.R
\name{report.glmmadmb}
\alias{report.glmmadmb}
\title{Defunct functions in package \sQuote{clickR}}
\usage{
report.glmmadmb(...)
}
\description{
Creates a report table from a glmmadmb model. This function has been
removed from the package. Updated functions are available in \sQuote{repmod} package.
}
\details{
Report from generalized linear mixed model from ADMB
}
