% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_model.R
\name{report.glmnet}
\alias{report.glmnet}
\title{Report models from glmnet package}
\usage{
\method{report}{glmnet}(x, s, drop.zero = TRUE, file = NULL,
  type = "word", digits = 3, info = TRUE, ...)
}
\arguments{
\item{x}{A glmnet model object}

\item{s}{Value of lambda for estimating the coefficients}

\item{drop.zero}{Should zero coefficients be dropped?}

\item{file}{Name of the file to export the table}

\item{type}{Format of the file}

\item{digits}{Number of decimals}

\item{info}{If TRUE, include call in the exported table}

\item{...}{Further arguments passed to make_table}
}
\value{
A data frame with the report table
}
\description{
Creates a report table from models fitted by glmnet
}
