% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_model.R
\name{report.lm}
\alias{report.lm}
\title{Defunct functions in package \sQuote{clickR}}
\usage{
report.lm(...)
}
\description{
Creates a report table from a linear model. This function has been
removed from the package. Updated functions are available in \sQuote{repmod} package.
}
\details{
Report from linear model
}
