% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_model.R
\name{report.lqmm}
\alias{report.lqmm}
\title{Defunct functions in package \sQuote{clickR}}
\usage{
report.lqmm(...)
}
\description{
Creates a report table from a quantile mixed model. This function has been
removed from the package. Updated functions are available in \sQuote{repmod} package.
}
\details{
Report from quantile mixed model
}
