% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_model.R
\name{report.numeric}
\alias{report.numeric}
\title{Report from numeric variable}
\usage{
\method{report}{numeric}(x, ...)
}
\arguments{
\item{x}{A numeric variable}

\item{...}{Further arguments passed to make_table}
}
\value{
A data frame with the report table
}
\description{
Creates a report table
}
