% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_model.R
\name{report.rlm}
\alias{report.rlm}
\title{Defunct functions in package \sQuote{clickR}}
\usage{
report.rlm(...)
}
\description{
Creates a report table from a robust linear model. This function has been
removed from the package. Updated functions are available in \sQuote{repmod} package.
}
\details{
Report from robust linear model (rlm)
}
