% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{restore_changes}
\alias{restore_changes}
\title{Restore changes}
\usage{
restore_changes(x, var.names)
}
\arguments{
\item{x}{A data.frame}

\item{var.names}{Character vector with names of the variables to be restored}
}
\description{
Restores original values after using a fix function
}
\examples{
mydata<-data.frame(Dates1=c("25/06/1983", "25-08/2014", "2001/11/01", "2008-10-01"),
                   Dates2=c("01/01/85", "04/04/1982", "07/12-2016", NA),
                   Numeric1=rnorm(4))
mydata <- fix.dates(mydata)
mydata
mydata <- restore_changes(mydata, "Dates1")
mydata
}
