% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{search_scripts}
\alias{search_scripts}
\title{Search scripts}
\usage{
search_scripts(string, path = getwd(), recursive = TRUE)
}
\arguments{
\item{string}{Character string to search}

\item{path}{Character vector with the path name}

\item{recursive}{Logical. Should the search be recursive into subdirectories?}
}
\value{
A list with each element being one of the files containing the search string
}
\description{
Searches for strings in R script files
}
