% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_model.R
\name{set_noms}
\alias{set_noms}
\title{Deprecated functions in package \sQuote{clickR}}
\usage{
set_noms(x, args)
}
\arguments{
\item{x}{A flextable object}

\item{args}{A names list with the header names}
}
\value{
A flextable object with assigned header names
}
\description{
Internal function for make_word_table. This function will be
defunct in the next version. Updated functions are available in \sQuote{repmod} package.
}
\details{
Set header names for word tables
}
