% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{track_changes}
\alias{track_changes}
\title{track_changes}
\usage{
track_changes(x, subset)
}
\arguments{
\item{x}{A data.frame}

\item{subset}{Logical expression for subsetting the data.frame with the changes}
}
\description{
Gets a data.frame with all the changes performed by the different fix functions
}
\examples{
mydata<-data.frame(Dates1=c("25/06/1983", "25-08/2014", "2001/11/01", "2008-10-01"),
                   Dates2=c("01/01/85", "04/04/1982", "07/12-2016", NA),
                   Numeric1=rnorm(4))
mydata <- fix.dates(mydata)
mydata
track_changes(mydata)
}
