% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_quality.R
\name{ttrue}
\alias{ttrue}
\title{True TRUE}
\usage{
ttrue(x)
}
\arguments{
\item{x}{A logical vector}
}
\value{
A logical vector
}
\description{
Makes possible vectorized logical comparisons against NULL and NA values
}
