% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_functions.R
\name{v_df_changes}
\alias{v_df_changes}
\title{Internal function to track_changes}
\usage{
v_df_changes(x, y)
}
\arguments{
\item{x}{Original data.frame}

\item{y}{New data.frame}
}
\description{
Function to track_changes
}
