% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_model.R
\name{VarCorr}
\alias{VarCorr}
\title{Generic VarCorr function}
\usage{
VarCorr(x, sigma = 1, ...)
}
\arguments{
\item{x}{A model object}

\item{sigma}{Optional value used as a multiplier for the standard deviations}

\item{...}{Further arguments passed to VarrCorr methods}
}
\value{
A Variance-Covariance Matrix
}
\description{
Extract Variance-Covariance Matrix
}
