% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_quality.R
\name{workspace}
\alias{workspace}
\title{Explores global environment workspace}
\usage{
workspace(table = FALSE)
}
\arguments{
\item{table}{If TRUE a table with the frequencies of each type of object is given}
}
\value{
A list of object names by class or a table with frequencies if table = TRUE
}
\description{
Returns information regarding the different objects in global environment
}
\examples{
df1 <- data.frame(x=rnorm(10), y=rnorm(10, 1, 2))
df2 <- data.frame(x=rnorm(20), y=rnorm(20, 1, 2))
workspace(table=TRUE)  #Frequency table of the different object classes
workspace()  #All objects in the global object separated by class
}
