% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_quality.R
\name{workspace_sapply}
\alias{workspace_sapply}
\title{Applies a function over objects of a specific class}
\usage{
workspace_sapply(object_class, action = "summary")
}
\arguments{
\item{object_class}{Class of the objects where the function is to be applied}

\item{action}{Name of the function to apply}
}
\value{
Results of the function
}
\description{
Applies a function over all objects of a specific class in the global environment
}
\examples{
df1 <- data.frame(x=rnorm(10), y=rnorm(10, 1, 2))
df2 <- data.frame(x=rnorm(20), y=rnorm(20, 1, 2))
workspace_sapply("data.frame", "summary")  #Gives a summary of each data.frame
}
