% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_quality.R
\name{xscores}
\alias{xscores}
\title{Estimate sample scores}
\usage{
xscores(x, type = "z")
}
\arguments{
\item{x}{A vector}

\item{type}{'z' calculates standard normal scores, 'z-out' calculates standard normal scores excluding
each data point when computing the mean and the standard deviation, 't' calculates t scores,
'chisq' calculates chisquared scores, 'tukey' calculates scores based on the boxplot method,
'mad' calculates scores using median and mad instead of mean and sd.}
}
\description{
Calculates different scores to measure how much extreme are the different data points
}
\examples{
xscores(iris$Sepal.Length, type="z-out")
}
