% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_model.R
\name{coefplot}
\alias{coefplot}
\title{Defunct functions in package \sQuote{clickR}}
\usage{
coefplot(...)
}
\description{
Creates a plot of the coefficients of a model. This function has been
removed from the package. Updated functions are available in \sQuote{repmod} package.
}
\details{
Plot of the coefficients of a model
}
