% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{fix.factors}
\alias{fix.factors}
\title{Fix factors imported as numerics}
\usage{
fix.factors(x, k = 5, drop = TRUE)
}
\arguments{
\item{x}{A data.frame}

\item{k}{Maximum number of numeric values to be converted to factor}

\item{drop}{Drop similar levels?}
}
\description{
Fixes factors imported as numerics
}
\examples{
report(mtcars)
report(fix.factors(mtcars))
}
