% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{fxd}
\alias{fxd}
\title{Internal function to fix.dates}
\usage{
fxd(d, locale = "C", use.probs = TRUE)
}
\arguments{
\item{d}{A character vector}

\item{locale}{Locale to be used for month names}

\item{use.probs}{Solve ambiguities by similarity to the most frequent formats}
}
\description{
Function to format dates
}
