% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_model.R
\name{make_csv_table}
\alias{make_csv_table}
\title{Export a table to excel}
\usage{
make_csv_table(x, file)
}
\arguments{
\item{x}{A data.frame object}

\item{file}{Name of the file}
}
\value{
Creates a .csv file with the table
}
\description{
Exports a table to Excel
}
