% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_model.R
\name{make_table}
\alias{make_table}
\title{Make a table from report}
\usage{
make_table(x, file, type, font = "Arial", pointsize = 11,
  add.rownames = TRUE)
}
\arguments{
\item{x}{A data.frame object}

\item{file}{Name of the file}

\item{type}{Type of file}

\item{font}{Font type}

\item{pointsize}{Size of font}

\item{add.rownames}{Should rownames be added to the output?}
}
\value{
Creates a file with the table
}
\description{
Auxiliary function to create tables
}
