% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_model.R
\name{make_word_table}
\alias{make_word_table}
\title{Export a table to word}
\usage{
make_word_table(x, pointsize, font, file, add.rownames, info)
}
\arguments{
\item{x}{A data.frame object}

\item{pointsize}{Font size}

\item{font}{Font type}

\item{file}{Name of the file}

\item{add.rownames}{Should rownames be added to the output?}

\item{info}{Footer for the table}
}
\value{
Creates a word with the table
}
\description{
Exports a table to Word
}
