% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{mine.plot}
\alias{mine.plot}
\title{Mine plot}
\usage{
mine.plot(
  x,
  fun = is.na,
  spacing = 5,
  sort = F,
  show.x = TRUE,
  show.y = TRUE,
  ...
)
}
\arguments{
\item{x}{A data.frame}

\item{fun}{A function that evaluates a vector and returns a logical vector}

\item{spacing}{Numerical separation between lines at the y-axis}

\item{sort}{If TRUE, variables are sorted according to their results}

\item{show.x}{Should the x-axis be plotted?}

\item{show.y}{Should the y-axis be plotted?}

\item{...}{further arguments passed to order()}
}
\description{
Creates a heatmap-like plot for exploring the data
}
\examples{
mine.plot(airquality)   #Displays missing data
mine.plot(airquality, fun=outliers)   #Shows extreme values
}
