% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_functions.R
\name{nice_names}
\alias{nice_names}
\title{Nice names}
\usage{
nice_names(x, select = 1:ncol(x), tolower = TRUE, track = TRUE)
}
\arguments{
\item{x}{A data.frame}

\item{select}{Numeric vector with the positions (all by default) to be affected by the function}

\item{tolower}{Set all names to lower case?}

\item{track}{Track changes?}
}
\value{
The input data.frame \code{x} with the fixed names
}
\description{
Changes names of a data frame to ease work with them
}
\examples{
d <- data.frame('Variable 1'=NA, '\% Response'=NA, ' Variable     3'=NA,check.names=FALSE)
names(d)
names(nice_names(d))
}
