% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{outliers}
\alias{outliers}
\title{outliers}
\usage{
outliers(x)
}
\arguments{
\item{x}{A vector}
}
\description{
Function for detecting outliers based on the boxplot method
}
\examples{
outliers(iris$Petal.Length)
outliers(airquality$Ozone)
}
