% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_model.R
\name{plot.reportmodel}
\alias{plot.reportmodel}
\title{Defunct functions in package \sQuote{clickR}}
\usage{
\method{plot}{reportmodel}(x, ...)
}
\description{
Creates a coefplot from the reportmodel object. This function has been
removed from the package. Updated functions are available in \sQuote{repmod} package.
}
\details{
Coefplot for reportmodel objects
}
