% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefplot.R
\name{plot.reportmodel}
\alias{plot.reportmodel}
\title{Coefplot for reportmodel objects}
\usage{
\method{plot}{reportmodel}(x, ...)
}
\arguments{
\item{x}{A reportmodel object}

\item{...}{Further arguments passed to coefplot}
}
\description{
Creates a coefplot from the reportmodel object
}
\examples{
lm1 <- lm(Petal.Length ~ Sepal.Width + Species, data=iris)
a<-report(lm1)
par(mar=c(4, 10, 3, 2))
plot(a)   #Coefplot calling plot.reportmodel
}
