% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_model.R
\name{report.brmsfit}
\alias{report.brmsfit}
\title{Report models from brms package}
\usage{
\method{report}{brmsfit}(x, file = NULL, type = "word", digits = 3,
  info = TRUE, ...)
}
\arguments{
\item{x}{A brms model object}

\item{file}{Name of the file to export the table}

\item{type}{Format of the file}

\item{digits}{Number of decimals}

\item{info}{If TRUE, include call in the exported table}

\item{...}{Further arguments passed to make_table}
}
\value{
A data frame with the report table
}
\description{
Creates a report table from model fitted by brms
}
