% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_model.R
\name{report.clmm}
\alias{report.clmm}
\title{Defunct functions in package \sQuote{clickR}}
\usage{
report.clmm(...)
}
\description{
Creates a report table from an ordinal mixed model. This function has been
removed from the package. Updated functions are available in \sQuote{repmod} package.
}
\details{
Report from ordinal mixed model
}
