% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_model.R
\name{report.data.frame}
\alias{report.data.frame}
\title{Defunct functions in package \sQuote{clickR}}
\usage{
report.data.frame(...)
}
\description{
Creates a report table ready for publication. This function has been
removed from the package. Updated functions are available in \sQuote{repmod} package.
}
\details{
Report tables of summary data
}
