% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_model.R
\name{report.factor}
\alias{report.factor}
\title{Deprecated functions in package \sQuote{clickR}}
\usage{
\method{report}{factor}(x, ...)
}
\arguments{
\item{x}{A categorical variable}

\item{...}{Further arguments passed to make_table}
}
\value{
A data frame with the report table
}
\description{
Creates a report table. This function will be
defunct in the next version. Updated functions are available in \sQuote{repmod} package.
}
\details{
Report from categorical variable
}
