% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_model.R
\name{report.factor}
\alias{report.factor}
\title{Defunct functions in package \sQuote{clickR}}
\usage{
report.factor(...)
}
\description{
Creates a report table. This function has been
removed from the package. Updated functions are available in \sQuote{repmod} package.
}
\details{
Report from categorical variable
}
