% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_model.R
\name{report.glmerMod}
\alias{report.glmerMod}
\title{Defunct functions in package \sQuote{clickR}}
\usage{
report.glmerMod(...)
}
\description{
Creates a report table from a generalized linear mixed model. This function has been
removed from the package. Updated functions are available in \sQuote{repmod} package.
}
\details{
Report from generalized linear mixed model
}
