% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_model.R
\name{report.glmmadmb}
\alias{report.glmmadmb}
\title{Report from generalized linear mixed model from ADMB}
\usage{
\method{report}{glmmadmb}(x, file = NULL, type = "word", digits = 3,
  digitspvals = 3, info = TRUE, ...)
}
\arguments{
\item{x}{A generalized linear mixed model object (glmmabmb)}

\item{file}{Name of the file to export the table}

\item{type}{Format of the file}

\item{digits}{Number of decimals}

\item{digitspvals}{Number of decimals for p-values}

\item{info}{If TRUE, include call in the exported table}

\item{...}{Further arguments passed to make_table}
}
\value{
A data frame with the report table
}
\description{
Creates a report table from a glmmadmb model
}
