% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_model.R
\name{report.rlm}
\alias{report.rlm}
\title{Deprecated functions in package \sQuote{clickR}}
\usage{
\method{report}{rlm}(
  x,
  file = NULL,
  type = "word",
  digits = 3,
  digitspvals = 3,
  info = TRUE,
  print = TRUE,
  ...
)
}
\arguments{
\item{x}{A rlm object}

\item{file}{Name of the file to export the table}

\item{type}{Format of the file}

\item{digits}{Number of decimals}

\item{digitspvals}{Number of decimals for p-values}

\item{info}{If TRUE, include call in the exported table}

\item{print}{Should the report table be printed on screen?}

\item{...}{Further arguments passed to make_table}
}
\value{
A data frame with the report table
}
\description{
Creates a report table from a robust linear model. This function will be
defunct in the next version. Updated functions are available in \sQuote{repmod} package.
}
\details{
Report from robust linear model (rlm)
}
