% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_model.R
\name{rob.ci}
\alias{rob.ci}
\title{Defunct functions in package \sQuote{clickR}}
\usage{
rob.ci(...)
}
\arguments{
\item{...}{Not used}
}
\description{
Estimates confidence intervals for rlm models. This function has been
removed from the package. Updated functions are available in \sQuote{repmod} package.
}
\details{
Function to compute bootstrap confidence intervals for robust linear regression models
}
