% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_model.R
\name{rob.pvals}
\alias{rob.pvals}
\title{Function to compute p-values for robust linear regression models}
\usage{
rob.pvals(x)
}
\arguments{
\item{x}{A rlm object}
}
\value{
A vector of p-values
}
\description{
Estimates p-values for rlm models
}
