% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_model.R
\name{set_noms}
\alias{set_noms}
\title{Defunct functions in package \sQuote{clickR}}
\usage{
set_noms(...)
}
\value{
A flextable object with assigned header names
}
\description{
Internal function for make_word_table. This function has been
removed from the package. Updated functions are available in \sQuote{repmod} package.
}
\details{
Set header names for word tables
}
