% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_model.R
\name{set_noms}
\alias{set_noms}
\title{Set header names for word tables}
\usage{
set_noms(x, args = list(...))
}
\arguments{
\item{x}{A flextable object}

\item{args}{A names list with the header names}
}
\value{
A flextable object with assigned header names
}
\description{
Internal function for make_word_table
}
